♻️ Refactor: 자치구 스토어 코드 분리
